data generator refactor
