+++ packages/core-ui/js/generator-script.js
@@ -1,5 +1,5 @@
 import { ensureGridLibraryLoaded } from './gui_components/data-grid-editor/grid-library-loader.js';
-import { DataGeneratorPage } from './gui_components/generator/data-generator-page.js';
+import { DataGeneratorPage } from './gui_components/generator/index.js';
 import { faker } from '@faker-js/faker';
 import { initHelpTooltips } from './help/help-tooltips.js';
 import { initThemeToggle } from './gui_components/shared/theme-toggle.js';
