+++ packages/core-ui/js/gui_components/generator/controller/index.js
@@ -1,0 +1,10 @@
+export {
+  DataGeneratorPage,
+  buildRuleSpecFromSchemaRow,
+  extractLiteralValueFromRuleSpec,
+  extractRegexValueFromRuleSpec,
+  schemaRowsToSpec,
+  schemaRowsToSpecWithTokens,
+  validateSchemaRows,
+  normaliseFakerCommand,
+} from './data-generator-page-controller.js';
