+++ packages/core-ui/js/gui_components/generator/host/index.js
@@ -1,0 +1,6 @@
+export {
+  GENERATE_TO_FILE_HELP_URL,
+  renderDataGeneratorPageShell,
+  populateGeneratorFormatOptions,
+} from './data-generator-page-layout.js';
+export { bindDataGeneratorPageEvents, initializeDataGeneratorPageHost } from './data-generator-page-coordinator.js';
