+++ packages/core-ui/js/gui_components/generator/options/index.js
@@ -1,0 +1,13 @@
+export {
+  TEST_FRAMEWORK_GROUPS,
+  getCodeLanguageSubtasks,
+  getTestFrameworkFormats,
+  getTestFrameworkLabel,
+  getUnitTestLanguageSubtasks,
+  sanitizeUiOptionsForFormat,
+} from './options-catalog-adapter.js';
+export {
+  OPTION_UI_SCHEMA_BY_FORMAT,
+  createOptionsPanelsForParent,
+  getOutputFormatGroups,
+} from './options-ui-schema.js';
