+++ packages/core-ui/js/gui_components/generator/options/options-catalog-adapter.js
