+++ packages/core-ui/js/gui_components/generator/options/options-ui-schema.js
@@ -1,22 +1,22 @@
-import { CsvDelimitedOptions } from '../options_panels/options-csv-delimited-controls.js';
-import { DelimitedOptions } from '../options_panels/options-delimited-controls.js';
-import { AsciiTableOptionsPanel } from '../options_panels/options-ascii-table.js';
-import { MarkdownOptionsPanel } from '../options_panels/options-markdown-panel.js';
-import { JsonOptionsPanel } from '../options_panels/options-json-panel.js';
-import { JavaOptionsPanel } from '../options_panels/options-java-panel.js';
-import { JavascriptOptionsPanel } from '../options_panels/options-javascript-panel.js';
-import { PythonOptionsPanel } from '../options_panels/options-python-panel.js';
-import { PhpOptionsPanel } from '../options_panels/options-php-panel.js';
-import { RubyOptionsPanel } from '../options_panels/options-ruby-panel.js';
-import { KotlinOptionsPanel } from '../options_panels/options-kotlin-panel.js';
-import { CSharpOptionsPanel } from '../options_panels/options-csharp-panel.js';
-import { PerlOptionsPanel } from '../options_panels/options-perl-panel.js';
-import { TypeScriptOptionsPanel } from '../options_panels/options-typescript-panel.js';
-import { XmlOptionsPanel } from '../options_panels/options-xml-panel.js';
-import { SqlOptionsPanel } from '../options_panels/options-sql-panel.js';
-import { GherkinOptionsPanel } from '../options_panels/options-gherkin-panel.js';
-import { HtmlOptionsPanel } from '../options_panels/options-html-panel.js';
-import { TestFrameworkOptionsPanel } from '../options_panels/options-test-framework-panel.js';
+import { CsvDelimitedOptions } from '../../options_panels/options-csv-delimited-controls.js';
+import { DelimitedOptions } from '../../options_panels/options-delimited-controls.js';
+import { AsciiTableOptionsPanel } from '../../options_panels/options-ascii-table.js';
+import { MarkdownOptionsPanel } from '../../options_panels/options-markdown-panel.js';
+import { JsonOptionsPanel } from '../../options_panels/options-json-panel.js';
+import { JavaOptionsPanel } from '../../options_panels/options-java-panel.js';
+import { JavascriptOptionsPanel } from '../../options_panels/options-javascript-panel.js';
+import { PythonOptionsPanel } from '../../options_panels/options-python-panel.js';
+import { PhpOptionsPanel } from '../../options_panels/options-php-panel.js';
+import { RubyOptionsPanel } from '../../options_panels/options-ruby-panel.js';
+import { KotlinOptionsPanel } from '../../options_panels/options-kotlin-panel.js';
+import { CSharpOptionsPanel } from '../../options_panels/options-csharp-panel.js';
+import { PerlOptionsPanel } from '../../options_panels/options-perl-panel.js';
+import { TypeScriptOptionsPanel } from '../../options_panels/options-typescript-panel.js';
+import { XmlOptionsPanel } from '../../options_panels/options-xml-panel.js';
+import { SqlOptionsPanel } from '../../options_panels/options-sql-panel.js';
+import { GherkinOptionsPanel } from '../../options_panels/options-gherkin-panel.js';
+import { HtmlOptionsPanel } from '../../options_panels/options-html-panel.js';
+import { TestFrameworkOptionsPanel } from '../../options_panels/options-test-framework-panel.js';
 import { getTestFrameworkFormats, getTestFrameworkLabel } from './options-catalog-adapter.js';
 
 const OPTION_UI_SCHEMA_BY_FORMAT = {
