+++ packages/core-ui/js/gui_components/generator/schema/index.js
@@ -1,0 +1,9 @@
+export {
+  buildSchemaModeHelpHtml,
+  hideVisibleHelpTooltips,
+  updateSchemaEditModeView,
+  insertExampleSchema,
+  renderGeneratorSchemaRows,
+  handleGeneratorRowInputChange,
+  handleGeneratorRowButtonClick,
+} from './data-generator-schema-ui.js';
