+++ tests/integration/cross-surface-option-parity.test.js
@@ -7,7 +7,7 @@ import { normalizeAndValidateFormat, sanitizeCliOptionsForFormat } from '../../a
 import {
   getTestFrameworkFormats,
   sanitizeUiOptionsForFormat,
-} from '../../packages/core-ui/js/gui_components/generator/options-catalog-adapter.js';
+} from '../../packages/core-ui/js/gui_components/generator/options/index.js';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
